requirements.txt
fastapi
uvicorn
openai
elevenlabs
python-dotenv
langchain
elevenlabs==2.22.1
